[O2B-1490] Update filterOnID test case in overview tests
Changed the filter input and expected results for a test case with filtering on env IDs.

This is because test inputs following each other that result in the same number of rows in the table will cause incorrect functioning of the waitForTableLength whose job it is is to wait for the filter to be applied.